+++ lib/sqlite-store.js
@@ -4,9 +4,19 @@
 // TODO: add some methods to create the test structure
 
 const Sqlite = require('sqlite3')
-
 const _ = require('lodash')
+
 const { intern } = require('./intern')
+
+const {
+  ensure_id,
+  is_new,
+  escape_str,
+  maybejsonfield,
+  makeent,
+  makeentp
+} = intern
+
 const error = require('eraro')({package: 'seneca-sqlite-store'})
 
 const name = 'sqlite-store'
@@ -152,10 +162,10 @@ module.exports = function (options) {
 
     save: function (args, done) {
       const ent = args.ent
-      const new_ent_id = intern.ensure_id(ent)
+      const new_ent_id = ensure_id(ent)
       const upsert_fields = is_upsert(args)
 
-      if (intern.is_new(ent) && upsert_fields) {
+      if (is_new(ent) && upsert_fields) {
         return do_upsert(upsert_fields, ent, new_ent_id, done)
       }
 
@@ -300,8 +310,8 @@ module.exports = function (options) {
       const table = tablename(qent)
 
       const stm = {
-        text: 'SELECT * FROM ' + escapeStr(table) + ' WHERE id IN (' +
-          escapeStr(idsstm(q)) + ')',
+        text: 'SELECT * FROM ' + escape_str(table) + ' WHERE id IN (' +
+          escape_str(idsstm(q)) + ')',
 
         values: q
       }
@@ -335,7 +345,7 @@ module.exports = function (options) {
       const metastr = ' ' + metaquery(qent, q)
 
       const stm = {
-        text: 'SELECT * FROM ' + escapeStr(table) + escapeStr(wherestr) + escapeStr(metastr),
+        text: 'SELECT * FROM ' + escape_str(table) + escape_str(wherestr) + escape_str(metastr),
         values: values
       }
 
@@ -438,7 +448,7 @@ module.exports = function (options) {
       params.push(fieldPlaceholder)
     })
 
-    stm.text = 'INSERT INTO ' + escapeStr(table) + ' (' + escapeStr(fields) + ') values (' + escapeStr(params) + ')'
+    stm.text = 'INSERT INTO ' + escape_str(table) + ' (' + escape_str(fields) + ') values (' + escape_str(params) + ')'
     stm.values = values
 
     return stm
@@ -482,7 +492,7 @@ module.exports = function (options) {
 
     values.$id = ent.id
 
-    stm.text = 'UPDATE ' + escapeStr(table) + ' SET ' + escapeStr(params) + ' WHERE id = $id'
+    stm.text = 'UPDATE ' + escape_str(table) + ' SET ' + escape_str(params) + ' WHERE id = $id'
     stm.values = values
 
     return stm
@@ -496,7 +506,7 @@ module.exports = function (options) {
 
     // TODO: Tidy up.
     //
-    const cols = Object.keys(entp_with_new_id).map(col => escapeStr(col)).join(', ')
+    const cols = Object.keys(entp_with_new_id).map(col => escape_str(col)).join(', ')
 
     const values = Object.keys(entp_with_new_id).reduce(function (h, field) {
       const placeholder = '$' + field
@@ -505,18 +515,18 @@ module.exports = function (options) {
         const unsafe_value = entp_with_new_id[field]
 
         return typeof unsafe_value === 'string'
-          ? escapeStr(unsafe_value)
+          ? escape_str(unsafe_value)
           : unsafe_value
       })()
 
-      h[escapeStr(placeholder)] = safe_value
+      h[escape_str(placeholder)] = safe_value
       return h
     }, {})
 
     const values_placeholders = Object.keys(values).join(', ')
 
     const upsert_on_str = upsert_fields
-      .map(x => escapeStr(x))
+      .map(x => escape_str(x))
       .join(', ')
 
     const update_kvs = (function () {
@@ -526,13 +536,13 @@ module.exports = function (options) {
 
       return Object.keys(kvs).map(field => {
         const placeholder = '$' + field
-        return [escapeStr(field), escapeStr(placeholder)].join(' = ')
+        return [escape_str(field), escape_str(placeholder)].join(' = ')
       }).join(', ')
     })()
 
     const stm = {
       text:
-        `insert into ${escapeStr(table)} (${cols})\n` +
+        `insert into ${escape_str(table)} (${cols})\n` +
         `values (${values_placeholders})\n` +
         `on conflict (${upsert_on_str})\n` +
         `do update set ${update_kvs}`,
@@ -582,16 +592,16 @@ module.exports = function (options) {
 
       if (metastr) {
         return `
-          DELETE FROM ${escapeStr(table)}
+          DELETE FROM ${escape_str(table)}
           WHERE id IN (
-            SELECT id FROM ${escapeStr(table)}
-            ${escapeStr(wherestr)}
-            ${escapeStr(metastr)}
+            SELECT id FROM ${escape_str(table)}
+            ${escape_str(wherestr)}
+            ${escape_str(metastr)}
           )
         `
       }
 
-      return `DELETE FROM ${escapeStr(table)} ${escapeStr(wherestr)}`
+      return `DELETE FROM ${escape_str(table)} ${escape_str(wherestr)}`
     })()
 
     return stm
@@ -628,102 +638,6 @@ module.exports = function (options) {
     })
   }
 
-  function makeentp(ent) {
-    const entp = {}
-    const fields = ent.fields$()
-
-    fields.forEach(function (field) {
-      if (!_.isDate(ent[field]) && _.isObject(ent[field])) {
-        entp[field] = JSON.stringify(ent[field])
-      }
-      else {
-        entp[field] = ent[field]
-      }
-    })
-
-    return entp
-  }
-
-  function makeent(ent, row, schema) {
-    const entp = {}
-    const fields = ent.fields$()
-
-    if (!_.isUndefined(ent) && !_.isUndefined(row)) {
-      fields.forEach(function (field) {
-        if (!_.isUndefined(row[field])) {
-          if (schema[field] === 'timestamp') {
-            entp[field] = new Date(JSON.parse(row[field]))
-          }
-          else {
-            const json = maybejsonfield(field, row, schema)
-
-            if (json.isjson) {
-              entp[field] = json.parsed
-            }
-            else {
-              entp[field] = row[field]
-            }
-          }
-        }
-      })
-    }
-
-    return ent.make$(entp)
-  }
-
-  function maybejsonfield(field, row, schema) {
-    /* NOTE:
-    Source: https://www.json.org/json-en.html
-
-    > JSON is built on two structures:
-    >
-    > A collection of name/value pairs. In various languages, this is realized
-    > as an object, record, struct, dictionary, hash table, keyed list, or
-    > associative array.
-    > 
-    > An ordered list of values. In most languages, this is realized as an
-    > array, vector, list, or sequence.
-    */
-    const mayTryParse = maystorejson(field, row, schema) &&
-      'string' === typeof row[field] && (
-        mayBeArrayJson(row[field]) ||
-        mayBeObjectJson(row[field])
-      )
-
-    if (!mayTryParse) {
-      return { isjson: false }
-    }
-
-    try {
-      const parsed = JSON.parse(row[field])
-      return { isjson: true, parsed }
-    } catch (err) {
-      if (err instanceof SyntaxError) {
-        return { isjson: false }
-      }
-
-      throw err
-    }
-
-    function mayBeArrayJson(s) {
-      return /^\s*\[/.test(s)
-    }
-
-    function mayBeObjectJson(s) {
-      return /^\s*\{/.test(s)
-    }
-  }
-
-  function maystorejson(field, row, schema) {
-    const coltype = schema[field]
-    const rowvalue = row[field]
-
-    // QUESTION: What about the BLOB type?
-    //
-    return /^text$|^varchar\(\d*\)$|^varchar$/.test(coltype) &&
-      'string' === typeof rowvalue
-  }
-
   function loadbyid(ent_id, ent, done) {
     // TODO: The #load method is part of the public interface of the store.
     // As such, it's a bad practice to invoke a public function from a private
@@ -748,33 +662,3 @@ module.exports = function (options) {
   return {name: store.name, tag: meta.tag}
 }
 
-
-function escapeStr(input) {
-  const str = '' + input
-
-  return str.replace(/[\0\b\t\x08\x09\x1a\n\r"'\\\%]/g, function (char) {
-    switch (char) {
-      case '\0':
-        return '\\0'
-      case '\x08':
-        return '\\b'
-      case '\b':
-        return '\\b'
-      case '\x09':
-        return '\\t'
-      case '\t':
-        return '\\t'
-      case '\x1a':
-        return '\\z'
-      case '\n':
-        return '\\n'
-      case '\r':
-        return '\\r'
-      case '\"':
-      case "'":
-      case '\\':
-      case '%':
-        return '\\' + char
-    }
-  })
-}
